Sonar fixes and others things
Added new functions to tools.h:
The first function, "convertToSafeInteger", is a template function that takes in two types: "ReturnValue" and "ConvertValue", as well as a value of "convertValue" of type "ConvertValue". It converts this value to the "ReturnValue" type and ensures that it falls within the range of 0 and the maximum value that the "ReturnValue" type can hold. It then returns the converted and safe value.

The second function, "getEnumClassNumber", is also a template function that takes in a single type: "EnumClass", and a reference to a value of "convertValue" of type "EnumClass". It uses the "magic_enum" library to convert the enumeration value to its underlying integral value and returns it.

+++ src/creatures/creature.h
@@ -352,7 +352,7 @@ class Creature : virtual public Thing
 
 		void onDeath();
 		virtual uint64_t getGainedExperience(Creature* attacker) const;
-		void addDamagePoints(Creature* attacker, int32_t damagePoints);
+		void addDamagePoints(Creature* attacker, int64_t damagePoints);
 		bool hasBeenAttacked(uint32_t attackerId);
 
 		//combat event functions
@@ -492,7 +492,7 @@ class Creature : virtual public Thing
 			}
 		}
 		struct CountBlock_t {
-			int32_t total;
+			int64_t total;
 			int64_t ticks;
 		};
 		using CountMap = std::map<uint32_t, CountBlock_t>;
